Added short keyword.

+++ tests/source-test56-malloc.c
@@ -1,4 +1,4 @@
-extern int* malloc(size_t s);
+extern int* malloc(long int s);
 extern void free(int* f);
 int main()
 {
